Ensure non-null OpenCensus string attribute values
Handling the error case for an effect attempts to record the underlying
exception's message as well as stack trace as span attributes.
Some exception types such as `java.net.ConnectException` do not have an
exception message, which used to result in null values for the `error.message`
attribute value.

OpenCensus does not permit null values to be recorded as the value
for a string attribute.  This would result in the error stack trace
being polluted with an additional top-level OpenCensus constructor
exception, obscuring the underlying issue.