Add scalajs support
For many applications, the root of a trace starts in the browser on the client.
It is valuable to track the root back to the client application, which may make
several server side requests as part of the same user gesture.

In addition, the transit time for large payloads, as well as the scalajs
processing time to unpack and work with larger payloads, can have a significant
impact on overall performance from a user experience perspective.  As such, it
is extremely valuable to understand performance from a user/browser perspective
in order to find slow spots and hot spots in an application.

I have another PR that for log4cats, that adds scalajs logging allowing logs to
be sent to an HTTP4S endpoint.  Using natchez with log4cats client side will
allow tidy client side trace support, logging along side other natchez /
log4cats logs server side.

This PR does little more than move files around, modify the build to support
crossProject / scalajs for `core` and `log`.